chore(packages): moving to local ts-logger and apollo utils
