+++ package.json
@@ -9,7 +9,7 @@
     "test": "dotenv -- turbo run test",
     "test-integrations": "dotenv -- turbo run test-integrations",
     "synth": "dotenv -- turbo run synth",
-    "clean": "rm -rf .turbo node_modules packages/**/.turbo packages/**/dist packages/**/node_modules servers/**/logs servers/**/.turbo servers/**/dist servers/**/node_modules lambdas/**/logs lambdas/**/.turbo lambdas/**/dist lambdas/**/node_modules infrastructure/**/dist infrastructure/**/node_modules infrastructure/**/.turbo infrastructure/**/cdktf.out"
+    "clean": "rm -rf .turbo node_modules packages/**/.turbo packages/**/dist packages/**/logs packages/**/node_modules servers/**/logs servers/**/.turbo servers/**/dist servers/**/logs servers/**/node_modules lambdas/**/logs lambdas/**/.turbo lambdas/**/dist lambdas/**/node_modules infrastructure/**/dist infrastructure/**/node_modules infrastructure/**/.turbo infrastructure/**/cdktf.out"
   },
   "dependencies": {
     "dotenv-cli": "latest"
