+++ servers/user-api/package.json
@@ -12,30 +12,12 @@
     "format": "eslint --fix \"src/**/*.ts\""
   },
   "dependencies": {
-    "@apollo/server": "4.9.5",
-    "@apollo/subgraph": "2.5.6",
+    "@pocket-tools/apollo-utils": "workspace:*",
     "@aws-sdk/client-eventbridge": "3.451.0",
     "@aws-sdk/client-pinpoint": "3.451.0",
-    "@opentelemetry/exporter-trace-otlp-grpc": "0.45.1",
-    "@opentelemetry/exporter-trace-otlp-http": "0.45.1",
-    "@opentelemetry/id-generator-aws-xray": "1.2.1",
-    "@opentelemetry/instrumentation-aws-sdk": "0.36.2",
-    "@opentelemetry/instrumentation-dataloader": "0.5.3",
-    "@opentelemetry/instrumentation-express": "0.33.3",
-    "@opentelemetry/instrumentation-graphql": "0.36.0",
-    "@opentelemetry/instrumentation-http": "0.45.1",
-    "@opentelemetry/instrumentation-knex": "0.32.3",
-    "@opentelemetry/instrumentation-mysql2": "0.34.3",
-    "@opentelemetry/instrumentation-net": "0.32.3",
-    "@opentelemetry/propagator-aws-xray": "1.3.1",
-    "@opentelemetry/resource-detector-aws": "1.3.3",
-    "@opentelemetry/sdk-node": "0.45.1",
-    "@opentelemetry/sdk-trace-node": "1.18.1",
-    "@pocket-tools/apollo-utils": "3.4.2",
-    "@pocket-tools/ts-logger": "^1.3.0",
+    "@pocket-tools/ts-logger": "workspace:*",
     "@sentry/node": "7.80.1",
     "@sentry/tracing": "7.80.1",
-    "aws-sdk": "2.1497.0",
     "dataloader": "2.2.2",
     "express": "4.18.2",
     "graphql": "16.8.1",
@@ -46,9 +28,9 @@
     "underscore.string": "3.3.6"
   },
   "devDependencies": {
-    "@faker-js/faker": "8.3.1",
     "eslint-config-custom": "workspace:*",
     "tsconfig": "workspace:*",
+    "@faker-js/faker": "8.3.1",
     "@types/chai": "4.3.10",
     "@types/chai-datetime": "0.0.39",
     "@types/jest": "29.5.8",
