+++ servers/user-api/src/config/index.ts
@@ -80,11 +80,8 @@ export default {
   },
   serviceName,
   tracing: {
-    host: process.env.OTLP_COLLECTOR_HOST || 'otlpcollector',
-    graphQLDepth: 8,
-    samplingRatio: 0.2,
-    grpcDefaultPort: 4317,
-    httpDEfaultPort: 4318,
+    host: process.env.OTLP_COLLECTOR_HOST || 'localhost',
+    serviceName: serviceName,
   },
   secrets: {
     contactHash: process.env.CONTACT_HASH || 'abcdefghijklmnop',
