+++ servers/user-api/src/logger.ts
@@ -1,0 +1,2 @@
+import { Logger, setLogger } from '@pocket-tools/ts-logger';
+export const serverLogger: Logger = setLogger();
