+++ servers/user-api/src/main.ts
@@ -1,10 +1,15 @@
 //this must run before all imports and server start
 //so open-telemetry can patch all libraries that we use
-import { nodeSDKBuilder } from './tracing';
-import { serverLogger, startServer } from './apollo';
+import { nodeSDKBuilder } from '@pocket-tools/apollo-utils';
+import { startServer } from './apollo';
+import { serverLogger } from './logger';
 import config from './config';
 
-nodeSDKBuilder().then(async () => {
+nodeSDKBuilder({
+  host: config.tracing.host,
+  serviceName: config.tracing.serviceName,
+  release: config.sentry.release,
+}).then(async () => {
   const { url } = await startServer(config.app.port);
   serverLogger.info(
     `🚀 Public server ready at http://localhost:${config.app.port}${url}`,
